pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "todos-fastapi"
version = "0.1.0"
description = "TODOs API with FastAPI, SQLAlchemy, PostgreSQL, JWT, Alembic, Docker, and tests."
authors = [{name="Your Name", email="you@example.com"}]
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "python-multipart>=0.0.9",
  "pydantic[email]>=2.7", 
  "pydantic-settings>=2.3",
  "SQLAlchemy>=2.0.30",
  "asyncpg>=0.29.0",
  "alembic>=1.13.2",
  "passlib[bcrypt]>=1.7.4",
  "python-jose[cryptography]>=3.3.0",
  "httpx>=0.27.0",
  "loguru>=0.7.2",
  "psycopg2-binary>=2.9.9",   
  "python-dotenv>=1.0.1",
  "email-validator>=2.0.0"  # Añadido explícitamente
]

[tool.setuptools]
packages = { find = { where = ["."], exclude = ["alembic*", "tests*"] } }

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-q"
testpaths = ["tests"]
asyncio_mode = "auto"